requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
cryptography==36.0.2
enum34==1.1.10
glob2==0.7
h11==0.13.0
idna==3.3
iniconfig==1.1.1
jsonpickle==2.1.0
lxml==4.8.0
Mako==1.2.0
MarkupSafe==2.1.1
namedlist==1.8
outcome==1.1.0
packaging==21.3
parse==1.19.0
parse-type==0.6.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
pytest==7.1.1
pytest-allure-adaptor2==1.8.2
pytest-bdd==5.0.0
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0
